requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
structlog==24.1.0
click==8.1.7

# Data ingestion
scapy==2.5.0
nfstream==6.6.0
pyshark==0.6
pcapy-ng==1.0.7

# Stream processing
kafka-python==2.0.2
confluent-kafka==2.3.0
faust-streaming==0.10.4

# State and storage
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Machine learning
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4
imbalanced-learn==0.11.0
joblib==1.3.2

# Deep learning (optional, uncomment if using)
# tensorflow==2.15.0
# torch==2.1.2

# API and web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Cloud SDKs
boto3==1.34.15
google-cloud-compute==1.14.1
azure-mgmt-network==25.0.0
azure-identity==1.15.0

# SDN and network automation
ryu==4.34
netmiko==4.3.0
ncclient==0.6.13
p4runtime-shell==0.4.0

# Monitoring and observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
pytest-mock==3.12.0
coverage==7.4.0
tox==4.11.4

# Utilities
requests==2.31.0
python-multipart==0.0.6
jsonschema==4.21.1
colorama==0.4.6
tqdm==4.66.1
